chore(xgrammar): split earley/prune_tests.rs out (500-LoC cap)
The Tier-3a prune tests pushed earley/tests.rs to 577 LoC. Extracted
the prune_tests module to its own file via #[path]; module hierarchy
(and all super:: paths) unchanged. 803 tests green.
